+++ lib/nxt_pipeline/pipeline.rb
@@ -35,9 +35,9 @@ module NxtPipeline
       step_resolvers << block
     end
 
-    def set_default_constructor(name)
+    def set_default_constructor(default_constructor)
       raise_duplicate_default_constructor if default_constructor_name.present?
-      self.default_constructor_name = name
+      self.default_constructor_name = default_constructor
     end
 
     def raise_duplicate_default_constructor
@@ -46,8 +46,7 @@ module NxtPipeline
 
     def step(argument = nil, **opts, &block)
       constructor = if block_given?
-        # make type the :to_s of inline steps
-        # fall back to :inline if no type is given
+        # make type the :to_s of inline steps fall back to :inline if no type is given
         argument ||= :inline
         opts.reverse_merge!(to_s: argument)
         Constructor.new(:inline, **opts, &block)
@@ -66,39 +65,32 @@ module NxtPipeline
         end
       end
 
-      steps << Step.new(argument, constructor, steps.count, **opts)
+      register_step(argument, constructor, callbacks, **opts)
     end
 
-    def execute(**changeset, &block)
+    def execute(**change_set, &block)
       reset
 
       configure(&block) if block_given?
-      before_execute_callback.call(self, changeset) if before_execute_callback.respond_to?(:call)
+      callbacks.run(:before, :execution, change_set)
 
-      result = steps.inject(changeset) do |changeset, step|
-        execute_step(step, **changeset)
-      rescue StandardError => error
-        logger_for_error = logger
-
-        error.define_singleton_method :details do
-          OpenStruct.new(
-            changeset: changeset,
-            logger: logger_for_error,
-            step: step
-          )
+      result = callbacks.around :execution, change_set do
+        steps.inject(change_set) do |set, step|
+          execute_step(step, **set)
+        rescue StandardError => error
+          decorate_error_with_details(error, set, step, logger)
+          handle_error_of_step(error)
+          set
         end
-
-        callback = find_error_callback(error)
-        raise unless callback && callback.continue_after_error?
-        handle_step_error(error)
-        changeset
       end
 
-      after_execute_callback.call(self, result) if after_execute_callback.respond_to?(:call)
+      callbacks.run(:after, :execution, change_set)
       result
     rescue StandardError => error
       handle_step_error(error)
     end
+
+    alias_method :call, :execute
 
     def handle_step_error(error)
       log_step(current_step)
@@ -115,22 +107,40 @@ module NxtPipeline
 
     alias :on_error :on_errors
 
-    def before_execute(&callback)
-      self.before_execute_callback = callback
+    def before_step(&block)
+      callbacks.register([:before, :step], block)
     end
 
-    def after_execute(&callback)
-      self.after_execute_callback = callback
+    def after_step(&block)
+      callbacks.register([:after, :step], block)
+    end
+
+    def around_step(&block)
+      callbacks.register([:around, :step], block)
+    end
+
+    def before_execution(&block)
+      callbacks.register([:before, :execution], block)
+    end
+
+    def after_execution(&block)
+      callbacks.register([:after, :execution], block)
+    end
+
+    def around_execution(&block)
+      callbacks.register([:around, :execution], block)
     end
 
     private
 
+    def callbacks
+      @callbacks ||= NxtPipeline::Callbacks.new(pipeline: self)
+    end
+
     attr_reader :error_callbacks, :constructors, :step_resolvers
     attr_accessor :current_step,
                   :current_arg,
-                  :default_constructor_name,
-                  :before_execute_callback,
-                  :after_execute_callback
+                  :default_constructor_name
 
     def default_constructor
       return unless default_constructor_name
@@ -138,12 +148,12 @@ module NxtPipeline
       @default_constructor ||= constructors[default_constructor_name.to_sym]
     end
 
-    def execute_step(step, **changeset)
+    def execute_step(step, **change_set)
       self.current_step = step
-      self.current_arg = changeset
-      result = step.execute(**changeset)
+      self.current_arg = change_set
+      result = step.execute(**change_set)
       log_step(step)
-      result || changeset
+      result || change_set
     end
 
     def find_error_callback(error)
@@ -167,6 +177,31 @@ module NxtPipeline
 
     def default_step_resolvers
       [->(step_argument) { step_argument.is_a?(Symbol) && step_argument }]
+    end
+
+    def decorate_error_with_details(error, change_set, step, logger)
+      error.define_singleton_method :details do
+        OpenStruct.new(
+          change_set: change_set,
+          logger: logger,
+          step: step
+        )
+      end
+      error
+    end
+
+    def register_step(argument, constructor, callbacks, **opts)
+      steps << Step.new(argument, constructor, steps.count, self, callbacks, **opts)
+    end
+
+    def handle_error_of_step(error)
+      error_callback = find_error_callback(error)
+      raise error unless error_callback.present? && error_callback.continue_after_error?
+
+      log_step(current_step)
+      raise error unless error_callback.present?
+
+      error_callback.call(current_step, current_arg, error)
     end
   end
 end